[internal][DatePicker] Move useInterceptProps into module
